--- app/src/main/java/com/codesoom/assignment/application/ProductService.java
@@ -1,60 +1,0 @@
-package com.codesoom.assignment.application;
-
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.domain.ProductRepository;
-import com.codesoom.assignment.dto.ProductData;
-import com.codesoom.assignment.errors.ProductNotFoundException;
-import com.github.dozermapper.core.Mapper;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-import java.util.List;
-
-@Service
-@Transactional
-public class ProductService {
-    private final Mapper mapper;
-    private final ProductRepository productRepository;
-
-    public ProductService(
-            Mapper dozerMapper,
-            ProductRepository productRepository
-    ) {
-        this.mapper = dozerMapper;
-        this.productRepository = productRepository;
-    }
-
-    public List<Product> getProducts() {
-        return productRepository.findAll();
-    }
-
-    public Product getProduct(Long id) {
-        return findProduct(id);
-    }
-
-    public Product createProduct(ProductData productData) {
-        Product product = mapper.map(productData, Product.class);
-        return productRepository.save(product);
-    }
-
-    public Product updateProduct(Long id, ProductData productData) {
-        Product product = findProduct(id);
-
-        product.changeWith(mapper.map(productData, Product.class));
-
-        return product;
-    }
-
-    public Product deleteProduct(Long id) {
-        Product product = findProduct(id);
-
-        productRepository.delete(product);
-
-        return product;
-    }
-
-    private Product findProduct(Long id) {
-        return productRepository.findById(id)
-                .orElseThrow(() -> new ProductNotFoundException(id));
-    }
-}
