--- app/src/main/java/com/codesoom/assignment/application/UserService.java
@@ -1,54 +1,0 @@
-package com.codesoom.assignment.application;
-
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.domain.UserRepository;
-import com.codesoom.assignment.dto.UserModificationData;
-import com.codesoom.assignment.dto.UserRegistrationData;
-import com.codesoom.assignment.errors.UserEmailDuplicationException;
-import com.codesoom.assignment.errors.UserNotFoundException;
-import com.github.dozermapper.core.Mapper;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-
-@Service
-@Transactional
-public class UserService {
-    private final Mapper mapper;
-    private final UserRepository userRepository;
-
-    public UserService(Mapper dozerMapper, UserRepository userRepository) {
-        this.mapper = dozerMapper;
-        this.userRepository = userRepository;
-    }
-
-    public User registerUser(UserRegistrationData registrationData) {
-        String email = registrationData.getEmail();
-        if (userRepository.existsByEmail(email)) {
-            throw new UserEmailDuplicationException(email);
-        }
-
-        User user = mapper.map(registrationData, User.class);
-        return userRepository.save(user);
-    }
-
-    public User updateUser(Long id, UserModificationData modificationData) {
-        User user = findUser(id);
-
-        User source = mapper.map(modificationData, User.class);
-        user.changeWith(source);
-
-        return user;
-    }
-
-    public User deleteUser(Long id) {
-        User user = findUser(id);
-        user.destroy();
-        return user;
-    }
-
-    private User findUser(Long id) {
-        return userRepository.findByIdAndDeletedIsFalse(id)
-                .orElseThrow(() -> new UserNotFoundException(id));
-    }
-}
