+++ app/src/main/java/com/codesoom/assignment/common/exception/ErrorResponse.java
@@ -1,0 +1,13 @@
+package com.codesoom.assignment.common.exception;
+
+public class ErrorResponse {
+    private String message;
+
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
