+++ app/src/main/java/com/codesoom/assignment/common/exception/GlobalExceptionHandler.java
@@ -1,0 +1,32 @@
+package com.codesoom.assignment.common.exception;
+
+import com.codesoom.assignment.product.exception.ProductNotFoundException;
+import com.codesoom.assignment.user.exception.UserEmailDuplicationException;
+import com.codesoom.assignment.user.exception.UserNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseBody
+@ControllerAdvice
+public class GlobalExceptionHandler {
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(ProductNotFoundException.class)
+    public ErrorResponse handleProductNotFound() {
+        return new ErrorResponse("Product not found");
+    }
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(UserNotFoundException.class)
+    public ErrorResponse handleUserNotFound() {
+        return new ErrorResponse("User not found");
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(UserEmailDuplicationException.class)
+    public ErrorResponse handleUserEmailIsAlreadyExisted() {
+        return new ErrorResponse("User's email address is already existed");
+    }
+}
