--- app/src/main/java/com/codesoom/assignment/domain/UserRepository.java
@@ -1,15 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import java.util.Optional;
-
-public interface UserRepository {
-    User save(User user);
-
-    boolean existsByEmail(String email);
-
-    Optional<User> findById(Long id);
-
-    Optional<User> findByIdAndDeletedIsFalse(Long id);
-
-    Optional<User> findByEmail(String email);
-}
