--- app/src/main/java/com/codesoom/assignment/dto/ProductData.java
@@ -1,31 +1,0 @@
-package com.codesoom.assignment.dto;
-
-import com.github.dozermapper.core.Mapping;
-import lombok.*;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Setter
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProductData {
-    private Long id;
-
-    @NotBlank
-    @Mapping("name")
-    private String name;
-
-    @NotBlank
-    @Mapping("maker")
-    private String maker;
-
-    @NotNull
-    @Mapping("price")
-    private Integer price;
-
-    @Mapping("imageUrl")
-    private String imageUrl;
-}
