--- app/src/main/java/com/codesoom/assignment/dto/UserModificationData.java
@@ -1,25 +1,0 @@
-package com.codesoom.assignment.dto;
-
-import com.github.dozermapper.core.Mapping;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserModificationData {
-    @NotBlank
-    @Mapping("name")
-    private String name;
-
-    @NotBlank
-    @Size(min = 4, max = 1024)
-    @Mapping("password")
-    private String password;
-}
