--- app/src/main/java/com/codesoom/assignment/dto/UserResultData.java
@@ -1,16 +1,0 @@
-package com.codesoom.assignment.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class UserResultData {
-    private Long id;
-
-    private String email;
-
-    private String name;
-}
