--- app/src/main/java/com/codesoom/assignment/errors/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment.errors;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
-        super("Product not found: " + id);
-    }
-}
