--- app/src/main/java/com/codesoom/assignment/errors/UserEmailDuplicationException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment.errors;
-
-public class UserEmailDuplicationException extends RuntimeException {
-    public UserEmailDuplicationException(String email) {
-        super("User email is alreay existed: " + email);
-    }
-}
