--- app/src/main/java/com/codesoom/assignment/errors/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment.errors;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Long id) {
-        super("User not found: " + id);
-    }
-}
