+++ app/src/main/java/com/codesoom/assignment/product/exception/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.product.exception;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(Long id) {
+        super("Product not found: " + id);
+    }
+}
