+++ app/src/main/java/com/codesoom/assignment/user/exception/UserEmailDuplicationException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.user.exception;
+
+public class UserEmailDuplicationException extends RuntimeException {
+    public UserEmailDuplicationException(String email) {
+        super("User email is alreay existed: " + email);
+    }
+}
