+++ app/src/main/java/com/codesoom/assignment/user/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.user.exception;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(Long id) {
+        super("User not found: " + id);
+    }
+}
