+++ app/src/main/java/com/codesoom/assignment/user/infra/JpaUserRepository.java
@@ -1,0 +1,20 @@
+package com.codesoom.assignment.user.infra;
+
+import com.codesoom.assignment.user.domain.User;
+import com.codesoom.assignment.user.domain.UserRepository;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface JpaUserRepository
+        extends UserRepository, CrudRepository<User, Long> {
+    User save(User user);
+
+    boolean existsByEmail(String email);
+
+    Optional<User> findById(Long id);
+
+    Optional<User> findByIdAndDeletedIsFalse(Long id);
+
+    Optional<User> findByEmail(String email);
+}
