+++ app/src/main/java/com/codesoom/assignment/user/presentation/UserController.java
@@ -1,0 +1,52 @@
+package com.codesoom.assignment.user.presentation;
+
+import com.codesoom.assignment.user.application.UserService;
+import com.codesoom.assignment.user.domain.User;
+import com.codesoom.assignment.user.presentation.dto.UserModificationData;
+import com.codesoom.assignment.user.presentation.dto.UserRegistrationData;
+import com.codesoom.assignment.user.presentation.dto.UserResultData;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+
+@RestController
+@RequestMapping("/users")
+@CrossOrigin
+public class UserController {
+    private final UserService userService;
+
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    UserResultData create(@RequestBody @Valid UserRegistrationData registrationData) {
+        User user = userService.registerUser(registrationData);
+        return getUserResultData(user);
+    }
+
+    @PatchMapping("{id}")
+    UserResultData update(
+            @PathVariable Long id,
+            @RequestBody @Valid UserModificationData modificationData
+    ) {
+        User user = userService.updateUser(id, modificationData);
+        return getUserResultData(user);
+    }
+
+    @DeleteMapping("{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    void destroy(@PathVariable Long id) {
+        userService.deleteUser(id);
+    }
+
+    private UserResultData getUserResultData(User user) {
+        return UserResultData.builder()
+                .id(user.getId())
+                .email(user.getEmail())
+                .name(user.getName())
+                .build();
+    }
+}
