+++ app/src/main/java/com/codesoom/assignment/user/presentation/dto/UserResultData.java
@@ -1,0 +1,16 @@
+package com.codesoom.assignment.user.presentation.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class UserResultData {
+    private Long id;
+
+    private String email;
+
+    private String name;
+}
