+++ app/src/test/java/com/codesoom/assignment/HelloControllerTest.java
@@ -1,0 +1,14 @@
+package com.codesoom.assignment;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class HelloControllerTest {
+    @Test
+    void sayHello() {
+        HelloController controller = new HelloController();
+
+        assertThat(controller.sayHello()).isEqualTo("Hello, world!");
+    }
+}
